+++ r/src/RcppExports.cpp
@@ -1795,6 +1795,19 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// iterate_matrix_linear_residual_cpp
+SEXP iterate_matrix_linear_residual_cpp(SEXP matrix, Eigen::Map<Eigen::ArrayXXd> row_params, Eigen::Map<Eigen::ArrayXXd> col_params);
+RcppExport SEXP _BPCells_iterate_matrix_linear_residual_cpp(SEXP matrixSEXP, SEXP row_paramsSEXP, SEXP col_paramsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type matrix(matrixSEXP);
+    Rcpp::traits::input_parameter< Eigen::Map<Eigen::ArrayXXd> >::type row_params(row_paramsSEXP);
+    Rcpp::traits::input_parameter< Eigen::Map<Eigen::ArrayXXd> >::type col_params(col_paramsSEXP);
+    rcpp_result_gen = Rcpp::wrap(iterate_matrix_linear_residual_cpp(matrix, row_params, col_params));
+    return rcpp_result_gen;
+END_RCPP
+}
 // iterate_matrix_scale_cpp
 SEXP iterate_matrix_scale_cpp(SEXP matrix, Eigen::Map<Eigen::ArrayXXd> row_scale, Eigen::Map<Eigen::ArrayXXd> col_scale);
 RcppExport SEXP _BPCells_iterate_matrix_scale_cpp(SEXP matrixSEXP, SEXP row_scaleSEXP, SEXP col_scaleSEXP) {
@@ -2594,6 +2607,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_BPCells_iterate_matrix_sctransform_pearson_transpose_cpp", (DL_FUNC) &_BPCells_iterate_matrix_sctransform_pearson_transpose_cpp, 4},
     {"_BPCells_iterate_matrix_sctransform_pearson_simd_cpp", (DL_FUNC) &_BPCells_iterate_matrix_sctransform_pearson_simd_cpp, 4},
     {"_BPCells_iterate_matrix_sctransform_pearson_transpose_simd_cpp", (DL_FUNC) &_BPCells_iterate_matrix_sctransform_pearson_transpose_simd_cpp, 4},
+    {"_BPCells_iterate_matrix_linear_residual_cpp", (DL_FUNC) &_BPCells_iterate_matrix_linear_residual_cpp, 3},
     {"_BPCells_iterate_matrix_scale_cpp", (DL_FUNC) &_BPCells_iterate_matrix_scale_cpp, 3},
     {"_BPCells_iterate_matrix_row_shift_cpp", (DL_FUNC) &_BPCells_iterate_matrix_row_shift_cpp, 2},
     {"_BPCells_iterate_matrix_col_shift_cpp", (DL_FUNC) &_BPCells_iterate_matrix_col_shift_cpp, 2},