+++ r/src/bpcells-cpp/matrixTransforms/LinearResidual.h
@@ -1,0 +1,38 @@
+// Copyright 2024 BPCells contributors
+// 
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
+
+#pragma once
+
+#include <atomic>
+#include "MatrixTransform.h"
+
+namespace BPCells {
+
+// Get linear residuals using QR decomposition
+class LinearResidual : public MatrixTransformDense {
+public:
+    using MatrixTransformDense::MatrixTransformDense;
+
+    bool loadZeroSubtracted(MatrixLoader<double> &loader) override;
+    void loadZero(double *values, uint32_t count, uint32_t start_row, uint32_t col) override;
+
+    Eigen::MatrixXd denseMultiplyRight(
+        const Eigen::Map<Eigen::MatrixXd> B, std::atomic<bool> *user_interrupt = NULL
+    ) override;
+    Eigen::MatrixXd denseMultiplyLeft(
+        const Eigen::Map<Eigen::MatrixXd> B, std::atomic<bool> *user_interrupt = NULL
+    ) override;
+    Eigen::VectorXd vecMultiplyRight(
+        const Eigen::Map<Eigen::VectorXd> v, std::atomic<bool> *user_interrupt = NULL
+    ) override;
+    Eigen::VectorXd vecMultiplyLeft(
+        const Eigen::Map<Eigen::VectorXd> v, std::atomic<bool> *user_interrupt = NULL
+    ) override;
+};
+
+} // namespace BPCells